+++ src/main/java/backend/like_house/domain/post/controller/PostController.java
@@ -188,24 +188,5 @@ public class PostController {
         postCommandService.togglePostAlarm(user, postId, enable);
         return ApiResponse.onSuccess(null);
     }
-
-    @PutMapping("/{commentId}/comment-alarm")
-    @Operation(summary = "댓글 알림 끄기/켜기 API", description = "사용자가 특정 댓글 알림을 끄거나 켜는 API입니다.")
-    @ApiResponses({
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON200", description = "OK, 성공"),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON400", description = "잘못된 요청입니다.")
-    })
-    @Parameters({
-            @Parameter(name = "commentId", description = "댓글의 ID, path variable 입니다."),
-            @Parameter(name = "enable", description = "댓글 알림 활성화 여부, query parameter 입니다.")
-    })
-    public ApiResponse<Void> toggleCommentAlarm(
-            @PathVariable Long commentId,
-            @Parameter(hidden = true) @LoginUser User user,
-            @RequestParam Boolean enable
-    ) {
-        postCommandService.toggleCommentAlarm(user, commentId, enable);
-        return ApiResponse.onSuccess(null);
-    }
 }
