+++ src/main/java/backend/like_house/domain/post/entity/Post.java
@@ -44,4 +44,8 @@ public class Post extends BaseEntity {
     public void setContent(String content) {
         this.content = content;
     }
+
+    public void setPostAlarm(Boolean postAlarm) {
+        this.postAlarm = postAlarm;
+    }
 }
