+++ src/main/java/backend/like_house/domain/post/service/CommentCommandService.java
@@ -8,4 +8,5 @@ public interface CommentCommandService {
     CreateCommentResponse createComment(User user, CreateCommentRequest request);
     CreateCommentResponse updateComment(User user, Long commentId, UpdateCommentRequest request);
     void deleteComment(User user, Long commentId);
+    void toggleCommentAlarm(User user, Long commentId, Boolean enable);
 }
