+++ src/main/java/backend/like_house/domain/post/service/PostCommandService.java
@@ -14,6 +14,5 @@ public interface PostCommandService {
     void likePost(User user, Long postId);
     void unlikePost(User user, Long postId);
     void togglePostAlarm(User user, Long postId, Boolean enable);
-    void toggleCommentAlarm(User user, Long commentId, Boolean enable);
 }
